requirements.txt
Flask==2.3.2
flask-sqlalchemy==3.0.3
flask-login==0.6.3
psycopg2-binary>=2.9.7
gunicorn==20.1.0

